+++ src/main/java/com/muriloCruz/ItGames/dto/post/PostSavedDto.java
@@ -31,7 +31,7 @@ public class PostSavedDto {
 
 	@Positive(message = "Game ID must be greater than 0")
 	@NotNull(message = "Game ID is required")
-	private Integer gameId;
+	private Long gameId;
 
 	@NotBlank(message = "Image URL is required")
 	private String imageUrl;
